refactor: extract pattern matcher utils to file
Move `hasGlob`, `globToRegex`, `escapeRegex`, and `makePatternMatcher`
into a shared utility file. Simplify the `testRunner`.

Related-Task: INTER-1466
